requirements.txt
altair==5.3.0
altgraph==0.17.4
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
deta==1.2.0
docutils==0.21.2
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_metadata==8.0.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==25.2.1
kiwisolver==1.4.5
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
more-itertools==10.3.0
nh3==0.2.17
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pkginfo==1.10.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
readme_renderer==43.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.18.1
setuptools==70.1.1
six==1.16.0
smmap==5.0.1
sql==2022.4.0
streamlit==1.36.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.13
tenacity==8.4.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
twine==5.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wheel==0.43.0
zipp==3.19.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.31
